flaskbook_api/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.5
opencv-python==4.6.0.66
Pillow==9.3.0
requests==2.28.1
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
